+++ service/controller/control_plane.go
@@ -38,7 +38,6 @@ func NewControlPlane(config ControlPlaneConfig) (*ControlPlane, error) {
 	var operatorkitController *controller.Controller
 	{
 		c := controller.Config{
-			CRD:          infrastructurev1alpha2.NewAWSControlPlaneCRD(),
 			K8sClient:    config.K8sClient,
 			Logger:       config.Logger,
 			ResourceSets: resourceSets,