+++ service/controller/control_plane_drainer_resource_set.go
@@ -10,36 +10,52 @@ import (
 	"github.com/giantswarm/operatorkit/resource"
 	"github.com/giantswarm/operatorkit/resource/wrapper/metricsresource"
 	"github.com/giantswarm/operatorkit/resource/wrapper/retryresource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 
 	"github.com/giantswarm/aws-operator/client/aws"
+	"github.com/giantswarm/aws-operator/pkg/label"
 	"github.com/giantswarm/aws-operator/pkg/project"
 	"github.com/giantswarm/aws-operator/service/controller/controllercontext"
 	"github.com/giantswarm/aws-operator/service/controller/key"
+	"github.com/giantswarm/aws-operator/service/controller/resource/asgname"
 	"github.com/giantswarm/aws-operator/service/controller/resource/asgstatus"
 	"github.com/giantswarm/aws-operator/service/controller/resource/awsclient"
-	"github.com/giantswarm/aws-operator/service/controller/resource/drainer"
-	"github.com/giantswarm/aws-operator/service/controller/resource/drainfinisher"
+	"github.com/giantswarm/aws-operator/service/controller/resource/drainerfinalizer"
+	"github.com/giantswarm/aws-operator/service/controller/resource/drainerinitializer"
 )
 
-type drainerResourceSetConfig struct {
+type controlPlaneDrainerResourceSetConfig struct {
 	G8sClient versioned.Interface
 	K8sClient kubernetes.Interface
 	Logger    micrologger.Logger
 
-	HostAWSConfig  aws.Config
-	ProjectName    string
-	Route53Enabled bool
+	HostAWSConfig aws.Config
+	ProjectName   string
 }
 
-func newDrainerResourceSet(config drainerResourceSetConfig) (*controller.ResourceSet, error) {
+func newControlPlaneDrainerResourceSet(config controlPlaneDrainerResourceSetConfig) (*controller.ResourceSet, error) {
 	var err error
+
+	var asgNameResource resource.Interface
+	{
+		c := asgname.Config{
+			Logger: config.Logger,
+
+			TagKey:       key.TagControlPlane,
+			TagValueFunc: key.ControlPlaneID,
+		}
+
+		asgNameResource, err = asgname.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
 
 	var asgStatusResource resource.Interface
 	{
 		c := asgstatus.Config{
-			G8sClient: config.G8sClient,
-			Logger:    config.Logger,
+			Logger: config.Logger,
 		}
 
 		asgStatusResource, err = asgstatus.New(c)
@@ -51,11 +67,11 @@ func newDrainerResourceSet(config drainerResourceSetConfig) (*controller.Resourc
 	var awsClientResource resource.Interface
 	{
 		c := awsclient.Config{
-			K8sClient:     config.K8sClient,
-			Logger:        config.Logger,
-			ToClusterFunc: newMachineDeploymentToClusterFunc(config.G8sClient),
+			K8sClient: config.K8sClient,
+			Logger:    config.Logger,
 
-			CPAWSConfig: config.HostAWSConfig,
+			CPAWSConfig:   config.HostAWSConfig,
+			ToClusterFunc: newControlPlaneToClusterFunc(config.G8sClient),
 		}
 
 		awsClientResource, err = awsclient.New(c)
@@ -64,28 +80,33 @@ func newDrainerResourceSet(config drainerResourceSetConfig) (*controller.Resourc
 		}
 	}
 
-	var drainerResource resource.Interface
+	var drainerInitializerResource resource.Interface
 	{
-		c := drainer.ResourceConfig{
-			G8sClient:     config.G8sClient,
-			Logger:        config.Logger,
-			ToClusterFunc: newMachineDeploymentToClusterFunc(config.G8sClient),
+		c := drainerinitializer.ResourceConfig{
+			G8sClient: config.G8sClient,
+			Logger:    config.Logger,
+
+			LabelMapFunc:  controlPlaneDrainerLabelMapFunc,
+			ToClusterFunc: newControlPlaneToClusterFunc(config.G8sClient),
 		}
 
-		drainerResource, err = drainer.NewResource(c)
+		drainerInitializerResource, err = drainerinitializer.NewResource(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
 	}
 
-	var drainFinisherResource resource.Interface
+	var drainerFinalizerResource resource.Interface
 	{
-		c := drainfinisher.ResourceConfig{
+		c := drainerfinalizer.ResourceConfig{
 			G8sClient: config.G8sClient,
 			Logger:    config.Logger,
+
+			LabelMapFunc:      controlPlaneDrainerLabelMapFunc,
+			LifeCycleHookName: key.LifeCycleHookControlPlane,
 		}
 
-		drainFinisherResource, err = drainfinisher.NewResource(c)
+		drainerFinalizerResource, err = drainerfinalizer.NewResource(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -93,9 +114,10 @@ func newDrainerResourceSet(config drainerResourceSetConfig) (*controller.Resourc
 
 	resources := []resource.Interface{
 		awsClientResource,
+		asgNameResource,
 		asgStatusResource,
-		drainerResource,
-		drainFinisherResource,
+		drainerInitializerResource,
+		drainerFinalizerResource,
 	}
 
 	{
@@ -119,7 +141,7 @@ func newDrainerResourceSet(config drainerResourceSetConfig) (*controller.Resourc
 	}
 
 	handlesFunc := func(obj interface{}) bool {
-		cr, err := key.ToMachineDeployment(obj)
+		cr, err := key.ToControlPlane(obj)
 		if err != nil {
 			return false
 		}
@@ -151,4 +173,11 @@ func newDrainerResourceSet(config drainerResourceSetConfig) (*controller.Resourc
 	}
 
 	return resourceSet, nil
+}
+
+func controlPlaneDrainerLabelMapFunc(cr metav1.Object) map[string]string {
+	return map[string]string{
+		label.Cluster:      key.ClusterID(cr),
+		label.ControlPlane: key.ControlPlaneID(cr),
+	}
 }