+++ service/controller/key/cluster.go
@@ -1,7 +1,7 @@
 package key
 
 import (
-	"crypto/sha1"
+	"crypto/sha512"
 	"fmt"
 	"strconv"
 	"strings"
@@ -47,6 +47,7 @@ const (
 	TagCluster                 = "giantswarm.io/cluster"
 	TagClusterType             = "giantswarm.io/cluster-type"
 	TagClusterTypeControlPlane = "control-plane"
+	TagControlPlane            = "giantswarm.io/control-plane"
 	TagInstallation            = "giantswarm.io/installation"
 	TagMachineDeployment       = "giantswarm.io/machine-deployment"
 	TagOrganization            = "giantswarm.io/organization"
@@ -61,6 +62,11 @@ const (
 	StackTCCPI = "tccpi"
 	StackTCNP  = "tcnp"
 	StackTCNPF = "tcnpf"
+)
+
+const (
+	LifeCycleHookControlPlane = "ControlPlane"
+	LifeCycleHookNodePool     = "NodePool"
 )
 
 const (
@@ -251,8 +257,11 @@ func ensureLabel(labels string, key string, value string) string {
 func getResourcenameWithTimeHash(prefix string, cluster infrastructurev1alpha2.AWSCluster, t time.Time) string {
 	id := strings.Replace(ClusterID(&cluster), "-", "", -1)
 
-	h := sha1.New()
-	h.Write([]byte(strconv.FormatInt(t.UnixNano(), 10)))
+	h := sha512.New()
+	_, err := h.Write([]byte(strconv.FormatInt(t.UnixNano(), 10)))
+	if err != nil {
+		panic(microerror.JSON(err))
+	}
 	timeHash := fmt.Sprintf("%x", h.Sum(nil))[0:5]
 
 	upperTimeHash := strings.ToUpper(timeHash)