+++ service/controller/resource/asgstatus/resource.go
@@ -6,14 +6,10 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/autoscaling"
-	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/giantswarm/apiextensions/pkg/clientset/versioned"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 
-	"github.com/giantswarm/aws-operator/pkg/awstags"
 	"github.com/giantswarm/aws-operator/service/controller/controllercontext"
-	"github.com/giantswarm/aws-operator/service/controller/key"
 )
 
 const (
@@ -21,26 +17,20 @@ const (
 )
 
 type Config struct {
-	G8sClient versioned.Interface
-	Logger    micrologger.Logger
+	Logger micrologger.Logger
 }
 
 type Resource struct {
-	g8sClient versioned.Interface
-	logger    micrologger.Logger
+	logger micrologger.Logger
 }
 
 func New(config Config) (*Resource, error) {
-	if config.G8sClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.G8sClient must not be empty", config)
-	}
 	if config.Logger == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
 	}
 
 	r := &Resource{
-		g8sClient: config.G8sClient,
-		logger:    config.Logger,
+		logger: config.Logger,
 	}
 
 	return r, nil
@@ -51,10 +41,6 @@ func (r *Resource) Name() string {
 }
 
 func (r *Resource) ensure(ctx context.Context, obj interface{}) error {
-	cr, err := key.ToMachineDeployment(obj)
-	if err != nil {
-		return microerror.Mask(err)
-	}
 	cc, err := controllercontext.FromContext(ctx)
 	if err != nil {
 		return microerror.Mask(err)
@@ -62,68 +48,22 @@ func (r *Resource) ensure(ctx context.Context, obj interface{}) error {
 
 	var asgName string
 	{
-		i := &ec2.DescribeInstancesInput{
-			Filters: []*ec2.Filter{
-				{
-					Name: aws.String(fmt.Sprintf("tag:%s", key.TagCluster)),
-					Values: []*string{
-						aws.String(key.ClusterID(&cr)),
-					},
-				},
-				{
-					Name: aws.String(fmt.Sprintf("tag:%s", key.TagMachineDeployment)),
-					Values: []*string{
-						aws.String(key.MachineDeploymentID(&cr)),
-					},
-				},
-				{
-					Name: aws.String(fmt.Sprintf("tag:%s", key.TagStack)),
-					Values: []*string{
-						aws.String(key.StackTCNP),
-					},
-				},
-				{
-					Name: aws.String("instance-state-name"),
-					Values: []*string{
-						aws.String(ec2.InstanceStateNamePending),
-						aws.String(ec2.InstanceStateNameRunning),
-						aws.String(ec2.InstanceStateNameStopped),
-						aws.String(ec2.InstanceStateNameStopping),
-					},
-				},
-			},
-		}
-
-		o, err := cc.Client.TenantCluster.AWS.EC2.DescribeInstances(i)
-		if IsNotFound(err) {
-			r.logger.LogCtx(ctx, "level", "debug", "message", "worker asg not available yet")
-			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
-			return nil
-		} else if err != nil {
-			return microerror.Mask(err)
-		}
-
-		if len(o.Reservations) == 0 {
-			r.logger.LogCtx(ctx, "level", "debug", "message", "worker asg not available yet")
-			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
-			return nil
-		}
-		if len(o.Reservations[0].Instances) == 0 {
-			r.logger.LogCtx(ctx, "level", "debug", "message", "worker asg not available yet")
+		if cc.Status.TenantCluster.ASG.Name == "" {
+			r.logger.LogCtx(ctx, "level", "debug", "message", "auto scaling group name not available yet")
 			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
 			return nil
 		}
 
-		asgName = awstags.ValueForKey(o.Reservations[0].Instances[0].Tags, "aws:autoscaling:groupName")
+		asgName = cc.Status.TenantCluster.ASG.Name
 	}
 
 	var asg *autoscaling.Group
 	{
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("finding ASG %#q", asgName))
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("finding auto scaling group %#q", asgName))
 
 		i := &autoscaling.DescribeAutoScalingGroupsInput{
 			AutoScalingGroupNames: []*string{
-				&asgName,
+				aws.String(asgName),
 			},
 		}
 
@@ -133,17 +73,17 @@ func (r *Resource) ensure(ctx context.Context, obj interface{}) error {
 		}
 
 		if len(o.AutoScalingGroups) != 1 {
-			return microerror.Maskf(executionFailedError, "there must be one item for ASG %#q", asgName)
+			return microerror.Maskf(executionFailedError, "there must be one item for auto scaling group %#q", asgName)
 		}
 		asg = o.AutoScalingGroups[0]
 
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("found ASG %#q", asgName))
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("found auto scaling group %#q", asgName))
 	}
 
 	var desiredCapacity int
 	{
 		if asg.DesiredCapacity == nil {
-			return microerror.Maskf(executionFailedError, "desired capacity must not be empty for ASG %#q", asgName)
+			return microerror.Maskf(executionFailedError, "desired capacity must not be empty for auto scaling group %#q", asgName)
 		}
 		desiredCapacity = int(*asg.DesiredCapacity)
 		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("desired capacity of %#q is %d", asgName, desiredCapacity))
@@ -152,7 +92,7 @@ func (r *Resource) ensure(ctx context.Context, obj interface{}) error {
 	var maxSize int
 	{
 		if asg.MaxSize == nil {
-			return microerror.Maskf(executionFailedError, "max size must not be empty for ASG %#q", asgName)
+			return microerror.Maskf(executionFailedError, "max size must not be empty for auto scaling group %#q", asgName)
 		}
 		maxSize = int(*asg.MaxSize)
 		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("max size of %#q is %d", asgName, maxSize))
@@ -161,7 +101,7 @@ func (r *Resource) ensure(ctx context.Context, obj interface{}) error {
 	var minSize int
 	{
 		if asg.MinSize == nil {
-			return microerror.Maskf(executionFailedError, "min size must not be empty for ASG %#q", asgName)
+			return microerror.Maskf(executionFailedError, "min size must not be empty for auto scaling group %#q", asgName)
 		}
 		minSize = int(*asg.MinSize)
 		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("min size of %#q is %d", asgName, minSize))
@@ -171,7 +111,6 @@ func (r *Resource) ensure(ctx context.Context, obj interface{}) error {
 		cc.Status.TenantCluster.ASG.DesiredCapacity = desiredCapacity
 		cc.Status.TenantCluster.ASG.MaxSize = maxSize
 		cc.Status.TenantCluster.ASG.MinSize = minSize
-		cc.Status.TenantCluster.ASG.Name = asgName
 	}
 
 	return nil